requirements.txt
arrow==0.8.0
atws==0.3.4
awsebcli==3.7.7
binaryornot==0.4.0
botocore==1.4.47
cement==2.8.2
chardet==2.3.0
click==6.6
colorama==0.3.7
cookiecutter==1.4.0
dj-database-url==0.3.0
dj-stripe==0.8.0
Django==1.9.7
django-appconf==1.0.2
django-braces==1.9.0
django-dotenv==1.4.1
django-model-utils==2.5.1
django-user-accounts==1.3.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
environ==1.0
future==0.15.2
Jinja2==2.8
jinja2-time==0.2.0
jmespath==0.9.0
jsonfield==1.0.3
MarkupSafe==0.23
pathspec==0.3.4
poyo==0.4.0
psycopg2==2.6.2
pycrypto==2.6.1
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
ssh==1.8.0
stripe==1.37.0
suds-jurko==0.6
texttable==0.8.4
websocket-client==0.37.0
whichcraft==0.4.0
whitenoise==3.2.1
